refactor: extract challenge logic from GameState into ChallengeManager (#350)
Move artist challenge and movie quiz challenge state, initialization,
and guess submission logic into a dedicated ChallengeManager class.
GameState delegates to the manager via properties and thin wrapper
methods, keeping its public interface fully backward-compatible.
